Restored functionality, improved endpoints. Code refactoring is still in progress.

+++ src/main/java/com/j2020/model/RevolutTokenRenewalResponse.java
@@ -33,7 +33,7 @@ public class RevolutTokenRenewalResponse implements TokenRenewalResponse {
         this.tokenType = tokenType;
     }
 
-    public int getSecondsUntilExpiring() {
+    public long getSecondsUntilExpiring() {
         return secondsUntilExpiring;
     }
